// client/src/api/models/NodeDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Notum nodes
 * 🚀🌙
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface NodeDto
 */
export interface NodeDto {
    /**
     * Node id
     * @type {number}
     * @memberof NodeDto
     */
    id: number;
    /**
     * Optional node's description
     * @type {string}
     * @memberof NodeDto
     */
    description?: string;
    /**
     * Parent ID (root node doesn't have parent)
     * @type {number}
     * @memberof NodeDto
     */
    parentId?: number;
    /**
     * Optional branch color
     * @type {string}
     * @memberof NodeDto
     */
    color?: string;
}

/**
 * Check if a given object implements the NodeDto interface.
 */
export function instanceOfNodeDto(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;

    return isInstance;
}

export function NodeDtoFromJSON(json: any): NodeDto {
    return NodeDtoFromJSONTyped(json, false);
}

export function NodeDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NodeDto {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'description': !exists(json, 'description') ? undefined : json['description'],
        'parentId': !exists(json, 'parentId') ? undefined : json['parentId'],
        'color': !exists(json, 'color') ? undefined : json['color'],
    };
}

export function NodeDtoToJSON(value?: NodeDto | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'description': value.description,
        'parentId': value.parentId,
        'color': value.color,
    };
}